bscan_spi: add script for Vivado XC7A fpgas
The build script uses Vivado to generate an indirect SPI flashing bitstream for a given part and package. The bitstream is placed in the bscan_spi folder and carries the name <part><package>.bit

Modify build.sh to change the part and package parameters. Ensure the required constraints (XDC) file is present.